backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
numpy>=1.26.0
opencv-python>=4.10.0
Pillow>=11.0.0
tensorflow>=2.18.0
ultralytics>=8.3.0
python-multipart>=0.0.19
